Remove unnecessary convertion.

+++ plugins/outputs/opentsdb/opentsdb.go
@@ -218,7 +218,7 @@ func buildValue(v interface{}) (string, error) {
 	case uint64:
 		retv = UIntToString(p)
 	case float64:
-		retv = FloatToString(float64(p))
+		retv = FloatToString(p)
 	default:
 		return retv, fmt.Errorf("unexpected type %T with value %v for OpenTSDB", v, v)
 	}
